수정: menu / loginBar 컴포넌트화
